# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API untuk manajemen keuangan pribadi (Cashflow).",
        "title": "Cashflow API Gin",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Bagas Rr",
            "url": "http://bagasrr.my.id"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/groups": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Membuat grup baru dengan anggota yang ditentukan.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Groups"
                ],
                "summary": "Create Group",
                "parameters": [
                    {
                        "description": "request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateGroupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.GroupResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
        "/groups/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mendapatkan detail grup berdasarkan ID, termasuk anggota dan informasi dompet.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Groups"
                ],
                "summary": "Get Group By ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Grup",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.GroupResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menghapus grup berdasarkan ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Groups"
                ],
                "summary": "Delete Group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Grup",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
        "/groups/{id}/remove-user": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menghapus pengguna dari grup.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Groups"
                ],
                "summary": "Remove User From Group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Grup",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.CreateGroupRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Kelompok untuk berbagi pengeluaran keluarga"
                },
                "member_ids": {
                    "description": "List user ID lain yg mau diajak (opsional)",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string",
                    "example": "Kelompok Keluarga"
                }
            }
        },
...
        "response.BaseResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "omitempty: sembunyikan jika nil"
                },
                "errors": {},
                "message": {
                    "type": "string",
                    "example": "success"
                },
                "status": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "response.CategoryResponse": {
            "type": "object",
            "properties": {
                "group_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "example": "Makanan"
                },
                "type": {
                    "type": "string",
                    "example": "EXPENSE"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "response.GroupMemberResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426655440000"
                },
                "role": {
                    "type": "string",
                    "example": "ADMIN"
                },
                "user_id": {
                    "description": "ID User aslinya",
                    "type": "string"
                },
                "username": {
                    "type": "string",
                    "example": "john_doe"
                }
            }
        },
        "response.GroupResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Kelompok untuk berbagi pengeluaran keluarga"
                },
                "id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426655440000"
                },
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.GroupMemberResponse"
                    }
                },
                "name": {
                    "type": "string",
                    "example": "Kelompok Keluarga"
                },
                "total_members": {
                    "type": "integer",
                    "example": 5
                },
                "wallet": {
                    "description": "Group pasti punya wallet",
                    "allOf": [
                        {
                            "$ref": "#/definitions/response.WalletResponse"
                        }
                    ]
                }
            }
        },
        "response.TransactionResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number",
                    "example": 500
                },
                "category": {
                    "$ref": "#/definitions/response.CategoryResponse"
                },
                "date": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2026-01-31T00:00:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "Gaji bulan Januari 2026"
                },
                "id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426655440000"
                },
                "title": {
                    "type": "string",
                    "example": "Gaji Bulanan"
                },
                "user": {
                    "$ref": "#/definitions/response.UserResponse"
                }
            }
        },
        "response.UserResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "john.doe@example.com"
                },
                "id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426655440000"
                },
                "user_role": {
                    "type": "string",
                    "example": "USER"
                },
                "username": {
                    "type": "string",
                    "example": "john_doe"
                },
                "wallets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.WalletResponse"
                    }
                }
            }
        },
        "response.WalletResponse": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "number",
                    "example": 1000
                },
                "group_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426655440000"
                },
                "name": {
                    "type": "string",
                    "example": "Tabungan"
                },
                "transaction_count": {
                    "type": "integer",
                    "example": 5
                },
                "transactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.TransactionResponse"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}